+++ src/file_manager/views.py
@@ -16,7 +16,8 @@ def index(request):
 
 @login_required
 def file_manager(request):
-    files = FileManager.objects.all()
+    logged_user_id = request.user.id
+    files = FileManager.objects.filter(user_id=logged_user_id)
     categories = FileType.objects.all()
     context = {
         'files': files,
@@ -39,6 +40,7 @@ def download(request, path):
 
 @login_required
 def upload(request):
+    logged_user_id = request.user.id
     context = {
         'form': UploadFile(),
     }
@@ -55,11 +57,19 @@ def upload(request):
         for category, extensions in files_types.items():
             if file_type in extensions:
                 file_category_id = FileType.objects.get(file_type=category).id
-                document = FileManager.objects.create(file_name=file, category_id_id=file_category_id)
+                document = FileManager.objects.create(
+                    user_id_id=logged_user_id,
+                    file_name=file,
+                    category_id_id=file_category_id,
+                )
                 document.save()
                 return redirect('file_manager')
         other_ctg_id = FileType.objects.get(file_type='Other').id
-        document = FileManager.objects.create(file_name=file, category_id_id=other_ctg_id)
+        document = FileManager.objects.create(
+            user_id_id=logged_user_id,
+            file_name=file,
+            category_id_id=other_ctg_id
+        )
         document.save()
         return redirect('file_manager')
     return render(request, 'pages/upload.html', context)
@@ -67,13 +77,15 @@ def upload(request):
 
 @login_required
 def delete_file(request, file_id):
-    FileManager.objects.get(pk=file_id).delete()
+    if FileManager.objects.get(id=file_id).user_id_id == request.user.id:
+        FileManager.objects.get(pk=file_id).delete()
     return redirect('file_manager')
 
 
 @login_required
 def show_by_category(request, category_id):
-    files = FileManager.objects.filter(category_id_id=category_id)
+    logged_user_id = request.user.id
+    files = FileManager.objects.filter(user_id=logged_user_id, category_id_id=category_id)
     categories = FileType.objects.all()
     context = {
         'files': files,
